Macros: move CellInfo, GridDimensions to collection
