+++ crates/kas-macros/src/collection.rs
@@ -7,11 +7,12 @@
 
 use proc_macro2::{Span, TokenStream as Toks};
 use quote::{quote, quote_spanned, ToTokens, TokenStreamExt};
-use syn::parse::{Parse, ParseStream, Result};
+use syn::parenthesized;
+use syn::parse::{Error, Parse, ParseStream, Result};
 use syn::punctuated::Punctuated;
 use syn::spanned::Spanned;
 use syn::token::Comma;
-use syn::{Expr, Ident, Lifetime, LitStr, Token};
+use syn::{Expr, Ident, Lifetime, LitInt, LitStr, Token};
 
 #[derive(Debug)]
 pub enum StorIdent {
@@ -54,6 +55,122 @@ impl NameGenerator {
         } else {
             Ok(self.next().into())
         }
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+pub struct CellInfo {
+    pub col: u32,
+    pub col_end: u32,
+    pub row: u32,
+    pub row_end: u32,
+}
+
+impl CellInfo {
+    pub fn new(col: u32, row: u32) -> Self {
+        CellInfo {
+            col,
+            col_end: col + 1,
+            row,
+            row_end: row + 1,
+        }
+    }
+}
+
+impl Parse for CellInfo {
+    fn parse(input: ParseStream) -> Result<Self> {
+        fn parse_end(input: ParseStream, start: u32) -> Result<u32> {
+            if input.parse::<Token![..=]>().is_ok() {
+                let lit = input.parse::<LitInt>()?;
+                let n: u32 = lit.base10_parse()?;
+                if n >= start {
+                    Ok(n + 1)
+                } else {
+                    Err(Error::new(lit.span(), format!("expected value >= {start}")))
+                }
+            } else if input.parse::<Token![..]>().is_ok() {
+                let plus = input.parse::<Token![+]>();
+                let lit = input.parse::<LitInt>()?;
+                let n: u32 = lit.base10_parse()?;
+
+                if plus.is_ok() {
+                    Ok(start + n)
+                } else if n > start {
+                    Ok(n)
+                } else {
+                    Err(Error::new(lit.span(), format!("expected value > {start}")))
+                }
+            } else {
+                Ok(start + 1)
+            }
+        }
+
+        let inner;
+        let _ = parenthesized!(inner in input);
+
+        let col = inner.parse::<LitInt>()?.base10_parse()?;
+        let col_end = parse_end(&inner, col)?;
+
+        let _ = inner.parse::<Token![,]>()?;
+
+        let row = inner.parse::<LitInt>()?.base10_parse()?;
+        let row_end = parse_end(&inner, row)?;
+
+        Ok(CellInfo {
+            row,
+            row_end,
+            col,
+            col_end,
+        })
+    }
+}
+
+impl ToTokens for CellInfo {
+    fn to_tokens(&self, toks: &mut Toks) {
+        let (col, col_end) = (self.col, self.col_end);
+        let (row, row_end) = (self.row, self.row_end);
+        toks.append_all(quote! {
+            ::kas::layout::GridCellInfo {
+                col: #col,
+                col_end: #col_end,
+                row: #row,
+                row_end: #row_end,
+            }
+        });
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct GridDimensions {
+    pub cols: u32,
+    col_spans: u32,
+    pub rows: u32,
+    row_spans: u32,
+}
+
+impl GridDimensions {
+    pub fn update(&mut self, cell: &CellInfo) {
+        self.cols = self.cols.max(cell.col_end);
+        if cell.col_end - cell.col > 1 {
+            self.col_spans += 1;
+        }
+        self.rows = self.rows.max(cell.row_end);
+        if cell.row_end - cell.row > 1 {
+            self.row_spans += 1;
+        }
+    }
+}
+
+impl ToTokens for GridDimensions {
+    fn to_tokens(&self, toks: &mut Toks) {
+        let (cols, rows) = (self.cols, self.rows);
+        let (col_spans, row_spans) = (self.col_spans, self.row_spans);
+        toks.append_all(quote! { ::kas::layout::GridDimensions {
+            cols: #cols,
+            col_spans: #col_spans,
+            rows: #rows,
+            row_spans: #row_spans,
+        } });
     }
 }
